Adding criptografy to password

+++ app.py
@@ -2,6 +2,7 @@ from flask import Flask, request, jsonify
 from models.user import User
 from database import db
 from flask_login import LoginManager, login_user, current_user, logout_user, login_required
+import bcrypt
 
 # Configuração para iniciar FLask
 app = Flask(__name__)
@@ -10,7 +11,6 @@ app.config['SECRET_KEY'] = "your_secret_key"
 # Endereço do Banco de Dados
 # SQLITE
 # app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
-app.config['SECRET_KEY'] = "your_secret_key"
 # MYSQL
 app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://root:admin123@127.0.0.1:3306/daily-diet'
 # Configuração para Banco de Dados
@@ -37,7 +37,7 @@ def login():
         # Login OK
         user = User.query.filter_by(username=username).first()
 
-        if user and user.password == password:
+        if user and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
             login_user(user)
             return jsonify({"message": "Autenticação realizada com sucesso"})
         
@@ -57,7 +57,8 @@ def create_user():
     password = data.get("password")
 
     if username and password:
-        user = User(username=username, password=password)
+        hashed_password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
+        user = User(username=username, password=hashed_password, role='user')
         db.session.add(user)
         db.session.commit()
         return jsonify({"message": "Usuário cadastrado com sucesso"})
@@ -87,6 +88,8 @@ def update_user(id_user):
     data = request.json
     user = User.query.get(id_user)
 
+    if id_user != current_user.id and current_user.role == "user":
+        return jsonify({"message": "Operação não permitida"}), 403
     if user and data.get("password"):
         user.password = data.get("password")
         db.session.commit()
@@ -94,18 +97,21 @@ def update_user(id_user):
 
     return jsonify({"message": "Usuário não encontrado"}), 404
 
+# Deleção de Usuários
 @app.route('/user/<int:id_user>', methods=["GET"])
 @login_required
 def delete_user(id_user):
     user = User.query.get(id_user)
 
+    if current_user.role != 'admin':
+        return jsonify({"message": "Operação não permitida"}), 403
     if id_user == current_user.id:
         return jsonify({"message": "Deleção não permitida"}), 403
 
     if user:
         db.session.delete(user)
         db.session.commit()
-        return jsonify({"message": f"Usiário {id_user} deletado com sucesso"})
+        return jsonify({"message": f"Usuário {id_user} deletado com sucesso"})
 
     return jsonify({"message": "Usuário não encontrado"}), 404
